Refactor json-related things to a separate file
Laying the ground work for the plugin architecture
